% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{alpha_to_e0_lt}
\alias{alpha_to_e0_lt}
\alias{alpha_to_e0_single}
\title{Calculate Life Expectancy at Birth Given Alpha}
\usage{
alpha_to_e0_lt(alpha, lx_std, pr_fem)

alpha_to_e0_single(alpha, lx_std, pr_fem)
}
\arguments{
\item{alpha}{Value from Brass logit model.
Length-1 numeric vector or rvec.}

\item{lx_std}{Standard schedule for 'lx'. Numeric vector.}

\item{pr_fem}{Proportion of births that are
female. Numeric scalar.}
}
\value{
A length-1 vector or rvec.
}
\description{
Life expectancy is for total population.
}
\details{
\code{alpha_to_e0_lt()} uses abridged life table ages
\code{alpha_to_e0_single()} using single year of age
}
