% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{e0_to_alpha_lt}
\alias{e0_to_alpha_lt}
\title{Calculate Alpha Given Single Value
for Life Expectancy at Birth - Life Table Ages}
\usage{
e0_to_alpha_lt(e0, lx_std, pr_fem)
}
\arguments{
\item{e0}{A length-1 numeric vector
or rvec}

\item{lx_std}{Numeric vector}

\item{pr_fem}{Proportion of births that are
female. Numeric scalar.}
}
\value{
A scalar
}
\description{
Calculate parameter from Brass logit model,
given (total) life expectancy,
for abridged life table
}
