% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{e0_to_alpha_lt_inner}
\alias{e0_to_alpha_lt_inner}
\title{Calculate Alpha Given Single Value
for Life Expectancy at Birth - Life Table Age groups}
\usage{
e0_to_alpha_lt_inner(e0, lx_std, pr_fem)
}
\arguments{
\item{e0}{A scalar}

\item{lx_std}{Numeric vector}

\item{pr_fem}{Proportion of births that are
female. Numeric scalar.}
}
\value{
A scalar
}
\description{
Calculate parameter from Brass logit model,
given (total) life expectancy
}
