% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit the TMB model}
\usage{
fit_model(popaccount, popn_obs)
}
\arguments{
\item{data_list}{A list of data}

\item{parameters_list}{A list of parameters}
}
\description{
Fit the TMB model
}
