% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{lx_to_Lx}
\alias{lx_to_Lx}
\title{Calculate 'Lx' from 'lx'}
\usage{
lx_to_Lx(lx, agetime_step)
}
\arguments{
\item{lx}{Lifetable column. Numeric vector or rvec.}

\item{age_time}{step}
}
\description{
Calculations differ according to
the value of \code{agetime_step}
}
