% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lx_west_7_complete}
\alias{lx_west_7_complete}
\title{'lx' Column from Abridged West Model Life Tables}
\format{
A data frame with 202 rows and the
following variables:
\itemize{
\item \code{sex} \code{"Female"}, and \code{"Male"}.
\item \code{age} Abridged life table age groups
up to \code{"100+"}.
\item \code{lx} Number of people still alive at start of
age interval.
}

A data frame with 202 rows and the
following variables:
\itemize{
\item \code{sex} \code{"Female"}, and \code{"Male"}.
\item \code{age} Single-year age groups, except for
the final age group, \code{"100+"}.
\item \code{lx} Number of people still alive at start of
age interval.
}
}
\source{
Calculated from
spreadsheet 'unpd_2011_mlt_130_2.5y_abridged.xlsx'
downloaded from the page "Model Life Tables"
on the United Nations Population Division
website on 19 January 2026.

Calculated from
spreadsheet 'unpd_2011_mlt_130_2.5y_complete.xlsx'
downloaded from the page "Model Life Tables"
on the United Nations Population Division
website on 19 January 2026.
}
\usage{
lx_west_7_complete

lx_west_7_complete
}
\description{
'lx' columns from "abridged"
life tables from the "West" family of Coale-Demeny
model life tables. The life tables have
life expectancies at birth
of 35 years for females and 32.5 years for males,
corresponding to "level 7" in the original
West model life table system.
An "abridged" life table uses age groups
\code{"0"}, \code{"1-4"}, \code{"5-9"}, ... rather than
single-year age groups.

'lx' columns from "complete" (ie 1-year age group)
life tables from the "West" family of Coale-Demeny
model life tables. The life tables have
life expectancies at birth
of 35 years for females and 32.5 years for males,
corresponding to "level 7" in the original
West model life table system.
}
\details{
The values in
\code{lx_west_7_complete} were not taken from the
original Coale-Demeny life tables, but instead from
a revised system constructed by the
United Nations Population Division. The revised
system extended the life table to age 130.
However, only values
up to age 100 are included here.

The values in
\code{lx_west_7_complete} were not taken from the
original Coale-Demeny life tables, but instead from
a revised system constructed by the
United Nations Population Division. The revised
system extended the life table to age 130.
However, only values
up to age 100 are included here.
}
\references{
The original Coale-Demeny system is
described in Coale A, Demeny P, and Vaughn B. 1983.
Regional model life tables and stable populations.
2nd ed. New York: Academic Press.

The revised system is described in
Li, N., and P. Gerland (2011).
Modifying the Lee-Carter method to
project mortality changes up to 2100.
In Annual Meeting of the Population Association of America.

The original Coale-Demeny system is
described in Coale A, Demeny P, and Vaughn B. 1983.
Regional model life tables and stable populations.
2nd ed. New York: Academic Press.

The revised system is described in
Li, N., and P. Gerland (2011).
Modifying the Lee-Carter method to
project mortality changes up to 2100.
In Annual Meeting of the Population Association of America.
}
\seealso{
\link{lx_west_7_complete} The "complete"
equivalent of \code{lx_west_7_abridged}.

\link{lx_west_7_abridged} The "abridged"
equivalent of \code{lx_west_7_complete},
using age groups \code{"0"}, \code{"1-4"}, \code{"5-9"}, ...
}
\keyword{datasets}
