% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popcounts-package.R
\docType{package}
\name{popcounts-package}
\alias{popcounts}
\alias{popcounts-package}
\title{Package 'popcounts'}
\description{
Estimate a Bayesian demographic account from imperfect data on population counts. Estimation is done using Template Model Builder.
}
\author{
\strong{Maintainer}: John Bryant \email{john@bayesiandemography.com}

Authors:
\itemize{
  \item Junni Zhang \email{junnizhang@163.com}
}

}
\keyword{internal}
